requirements.txt
tensorflow-gpu
numpy
scikit-image
scikit-learn
cupy
